Fix Bin out put and add notifications

+++ CORESI.Tools/TextAppender.cs
@@ -6,11 +6,13 @@ namespace CORESI.Tools
     {
         public static void Append(string fileName, string text, bool deleteFileIfExist = false)
         {
-            if (deleteFileIfExist)
+            if (deleteFileIfExist && File.Exists(fileName))
             {
                 File.Delete(fileName);
             }
 
+            var directory = Path.GetDirectoryName(fileName);
+            Directory.CreateDirectory(directory);
             using (var sr = File.AppendText(fileName))
             {
                 sr.WriteLine(text);
